Move the avatar menu to the floating menu bar if enabled
This makes the main screen look more balanced with both sides of the
view mode menu being accompanied by an addition item. One side has the
avatar menu, the other has the post composer button.

The avatar menu still needs styling to fit in with the rest of the menu.
Atm it is just the same circle as in the toolbar, which is slightly too
small for the floating menu.